+++ src/function/define.js
@@ -2,7 +2,8 @@ import { defineInstance } from '../core/function';
 import * as methods from './methods';
 
 defineInstance('debounce', methods.debounce);
-defineInstance('delay', methods.delay);
 defineInstance('memoize', methods.memoize);
 defineInstance('once', methods.once);
+defineInstance('setInterval', methods.setInterval);
+defineInstance('setTimeout', methods.setTimeout);
 defineInstance('throttle', methods.throttle);