+++ src/function/methods.js
@@ -1,5 +1,6 @@
 export { default as debounce } from './debounce';
-export { default as delay } from './delay';
 export { default as memoize } from './memoize';
 export { default as once } from './once';
+export { default as setInterval } from './setInterval';
+export { default as setTimeout } from './setTimeout';
 export { default as throttle } from './throttle';